requirements.txt
# base requirements for manyx project
Django==2.1.2
djangorestframework==3.9.0
django-filter==2.0.0
# installation of postgresql
# sudo apt-get install python-pip python-dev libpq-dev postgresql postgresql-contrib
psycopg2-binary==2.7.4
django-jalali-date==0.2.9
django-jalali==3.1.0
Pillow==5.3.0
# getting ip address for each client.
django-ipware==2.1.0
